+++ demos/session_automotive/src/main/java/androidx/media3/demo/session/automotive/AutomotiveService.kt
@@ -40,7 +40,7 @@ class AutomotiveService : DemoPlaybackService() {
     return exoPlayer
   }
 
-  override fun createLibrarySessionCallback(): MediaLibrarySession.Callback {
+  @OptIn(UnstableApi::class) override fun createLibrarySessionCallback(): MediaLibrarySession.Callback {
     return object : DemoMediaLibrarySessionCallback(this@AutomotiveService) {
 
       @OptIn(UnstableApi::class)