+++ libraries/session/src/test/java/androidx/media3/session/DefaultMediaNotificationProviderTest.java
@@ -48,7 +48,6 @@ import com.google.common.collect.ImmutableList;
 import com.google.common.util.concurrent.Futures;
 import com.google.common.util.concurrent.ListenableFuture;
 import com.google.common.util.concurrent.SettableFuture;
-import java.util.ArrayList;
 import java.util.List;
 import org.junit.Before;
 import org.junit.Test;
@@ -726,20 +725,8 @@ public class DefaultMediaNotificationProviderTest {
   public void createNotification_withCustomButtons_showsCustomButtonsWhereApplicable() {
     DefaultActionFactory defaultActionFactory =
         new DefaultActionFactory(Robolectric.setupService(TestService.class));
-    List<CommandButton> filteredMediaButtonPreferences = new ArrayList<>();
     DefaultMediaNotificationProvider defaultMediaNotificationProvider =
-        new DefaultMediaNotificationProvider(ApplicationProvider.getApplicationContext()) {
-          @Override
-          protected ImmutableList<CommandButton> getMediaButtons(
-              MediaSession session,
-              Commands playerCommands,
-              ImmutableList<CommandButton> mediaButtonPreferences,
-              boolean showPauseButton) {
-            filteredMediaButtonPreferences.addAll(mediaButtonPreferences);
-            return super.getMediaButtons(
-                session, playerCommands, mediaButtonPreferences, showPauseButton);
-          }
-        };
+        new DefaultMediaNotificationProvider(ApplicationProvider.getApplicationContext());
     MediaSession mediaSession =
         new MediaSession.Builder(
                 ApplicationProvider.getApplicationContext(),
@@ -764,13 +751,14 @@ public class DefaultMediaNotificationProviderTest {
             .setPlayerCommand(Player.COMMAND_PLAY_PAUSE)
             .build();
 
+//<<<<<<< HEAD
     defaultMediaNotificationProvider.createNotification(
         mediaSession,
         /* mediaButtonPreferences= */ ImmutableList.of(button1, button2, button3),
         defaultActionFactory,
         notification -> {
           /* Do nothing. */
-        });
+        });/*
 
 <<<<<<< HEAD
     assertThat(filteredMediaButtonPreferences).containsExactly(button2);
@@ -783,6 +771,20 @@ public class DefaultMediaNotificationProviderTest {
         .isEqualTo(context.getString(R.string.media3_controls_play_description));
     assertThat(mediaNotification.notification.actions[2].title.toString()).isEqualTo("button1");
 >>>>>>> 8b334e45a9 (Use standard commands in custom layout conversion)
+=======
+    MediaNotification mediaNotification =
+        defaultMediaNotificationProvider.createNotification(
+            mediaSession,
+            * mediaButtonPreferences= * ImmutableList.of(button1, button2, button3),
+            defaultActionFactory,
+            notification -> {});
+    mediaSession.release();
+
+    assertThat(mediaNotification.notification.actions).hasLength(2);
+    assertThat(mediaNotification.notification.actions[0].title.toString())
+        .isEqualTo(context.getString(R.string.media3_controls_play_description));
+    assertThat(mediaNotification.notification.actions[1].title.toString()).isEqualTo("button2");
+>>>>>>> d3e54fdff3 (Remove duplicate filtering)*/
   }
 
   @Test