+++ libraries/test_session_current/src/androidTest/java/androidx/media3/session/MediaSessionTest.java
@@ -64,6 +64,7 @@ import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.Future;
 import java.util.concurrent.TimeoutException;
+import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.atomic.AtomicReference;
 import org.junit.After;
@@ -387,15 +388,10 @@ public class MediaSessionTest {
   @Test
   public void creatingTwoSessionWithSameId() {
     String sessionId = "testSessionId";
-    MediaSession session =
-        new MediaSession.Builder(
-                context, new MockPlayer.Builder().setApplicationLooper(handler.getLooper()).build())
-            .setId(sessionId)
-            .build();
+    Player player = new MockPlayer.Builder().setApplicationLooper(handler.getLooper()).build();
+    MediaSession session = new MediaSession.Builder(context, player).setId(sessionId).build();
 
-    MediaSession.Builder builderWithSameId =
-        new MediaSession.Builder(
-            context, new MockPlayer.Builder().setApplicationLooper(handler.getLooper()).build());
+    MediaSession.Builder builderWithSameId = new MediaSession.Builder(context, player);
     try {
       builderWithSameId.setId(sessionId).build();
       assertWithMessage(
@@ -407,7 +403,8 @@ public class MediaSessionTest {
 
     session.release();
     // Creating a new session with ID of the closed session is okay.
-    MediaSession sessionWithSameId = builderWithSameId.build();
+    MediaSession sessionWithSameId =
+        new MediaSession.Builder(context, player).setId(sessionId).build();
     sessionWithSameId.release();
   }
 
@@ -482,6 +479,33 @@ public class MediaSessionTest {
 
     player.awaitMethodCalled(MockPlayer.METHOD_SEEK_TO, TIMEOUT_MS);
     assertThat(player.seekPositionMs).isEqualTo(testSeekPositionMs);
+  }
+
+  @Test
+  public void getPlatformToken_inApp_ensuresControllerInfoIsTrusted() throws Exception {
+    AtomicBoolean controllerIsTrusted = new AtomicBoolean();
+    MediaSession session =
+        sessionTestRule.ensureReleaseAfterTest(
+            new MediaSession.Builder(context, player)
+                .setId("getPlatformToken_inApp_ensuresControllerInfoIsTrusted")
+                .setCallback(
+                    new MediaSession.Callback() {
+                      @Override
+                      public MediaSession.ConnectionResult onConnect(
+                          MediaSession session, ControllerInfo controller) {
+                        controllerIsTrusted.set(controller.isTrusted());
+                        return MediaSession.Callback.super.onConnect(session, controller);
+                      }
+                    })
+                .build());
+    android.media.session.MediaSession.Token token = session.getPlatformToken();
+
+    android.media.session.MediaController platformController =
+        new android.media.session.MediaController(context, token);
+    platformController.getTransportControls().seekTo(100);
+    player.awaitMethodCalled(MockPlayer.METHOD_SEEK_TO, TIMEOUT_MS);
+
+    assertThat(controllerIsTrusted.get()).isTrue();
   }
 
   @Test