+++ RELEASENOTES.md
@@ -44,8 +44,32 @@
         describes.
 *   IMA extension:
 *   Session:
+<<<<<<< HEAD
+<<<<<<< HEAD
     *   Fix notification artwork icon scaling on Android Q
         ([#3084](https://github.com/androidx/media/issues/3084)).
+=======
+    *   Fix bug where 'ForegroundServiceStartNotAllowedException' wasn't
+        propagated across thread boundaries in case the application thread is
+        not the main thread
+        ([#2499](https://github.com/androidx/media/issues/2499)).
+    *   Fix bug where `COMMAND_SEEK_TO_MEDIA_ITEM` is not available in a
+        `MediaController` when connecting to a platform-only session that
+        supports seeking to other items but no playlist modifications.
+    *   Add support to override package name while creating MediaSession via a
+        new API `MediaSession.Builder#setPackageNameOverride`. The caller must
+        have `android.permission.OVERRIDE_MEDIA_SESSION_OWNER` permission to
+        override the package name of the session.
+    *   Fix issue where `MediaController` can't set media items on platform
+        media sessions that don't allow `PREPARE_FROM` actions.
+    *   Fix bug where own process wasn't marked with `Controller.isTrusted()` on
+        API 27 or before when connecting via platform controllers.
+>>>>>>> a90ff4a5be (Fix isTrusted for own app on API<28)
+=======
+    *   Workaround SysUI bug on API 29 where media session artwork in a
+        particular size can crash SysUI and force the device to reboot
+        ([#3118](https://github.com/androidx/media/issues/3118)).
+>>>>>>> 2d454f57a4 (Fix SysUI crash/device reboot issue on API 29 for specific bitmap size)
 *   UI:
     *   Hide playback position from the accessibility service to prevent
         continuous announcement update when `PlayerControlView` is visible.
@@ -231,8 +255,16 @@ This release includes the following changes since
         preferred orientation for linear ads. If you want to join the alpha
         testing group, please reach out to your account partner.
 *   Session:
+<<<<<<< HEAD
     *   Fix a bug where large arrays in `MediaMetadata.artworkData` can crash
         the app.
+=======
+    *   Fix issue where system UI button placement workaround negatively affects
+        other UI surface like Android Auto or manufacturers not needing the
+        workaround ([#3041]()).
+    *   Fix bug where large arrays in `MediaMetadata.artworkData` can crash the
+        app.
+>>>>>>> eeb5a2a398 (Fix workaround logic for system UI gaps)
     *   Use type int for `@Player.Command` argument in the `ActionFactory`
         interface ([#2933](https://github.com/androidx/media/issues/2933)).
     *   Make `MediaSessionService` and `MediaLibraryService` a